break out validateHostPort into its own file
