+++ v2/handler.js
@@ -38,6 +38,7 @@ var StreamingInResponse = require('../streaming_in_response');
 
 var v2 = require('./index');
 var errors = require('../errors');
+var validateHostPort = require('../lib/validate_host_port');
 
 var TCHANNEL_LANGUAGE = 'node';
 var TCHANNEL_LANGUAGE_VERSION = process.version.slice(1);
@@ -972,39 +973,4 @@ function InRequestOptions(
     self.connection = connection;
     self.hostPort = hostPort;
     self.tracer = tracer;
-}
-
-function stringIsValidNumber(numAsStr) {
-    var num = parseInt(numAsStr, 10);
-    return num.toString() === numAsStr;
-}
-
-function validateHostPort(hostPort) {
-    var parts = hostPort.split(':');
-    if (parts.length !== 2) {
-        return false;
-    }
-
-    var hostParts = parts[0].split('.');
-    if (hostParts.length !== 4) {
-        return false;
-    }
-
-    var i;
-    for (i = 0; i < 4; i++) {
-        if (!stringIsValidNumber(hostParts[i])) {
-            return false;
-        }
-    }
-
-    if (!stringIsValidNumber(parts[1])) {
-        return false;
-    }
-
-    var portNum = parseInt(parts[1], 10);
-
-    if (portNum < 0 || portNum > 65536) {
-        return false;
-    }
-    return true;
 }
